lsc-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
numpy
pillow
opencv-python-headless
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu